TypeScriptified `ghostLocals` middleware
