+++ ghost/core/core/server/web/parent/middleware/index.js
@@ -1,6 +1,6 @@
 module.exports = {
   emitEvents: require('./emit-events'),
-  ghostLocals: require('./ghost-locals'),
+  ghostLocals: require('./ghost-locals').ghostLocals,
   logRequest: require('./log-request'),
   queueRequest: require('./queue-request'),
   requestId: require('./request-id'),
